# BE_E_Learning/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Learning API",
        "description": "Tài liệu API cho hệ thống học trực tuyến",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/courses": {
            "get": {
                "tags": [
                    "Course"
                ],
                "summary": "Lấy danh sách khoá học",
                "operationId": "5dcadcedaa5505d09b15f882cd0c0cf7",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Tìm theo tiêu đề",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Lọc theo danh mục",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "instructor_id",
                        "in": "query",
                        "description": "Lọc theo giảng viên",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách khoá học"
                    }
                }
            },
            "post": {
                "tags": [
                    "Course"
                ],
                "summary": "Tạo khoá học mới",
                "operationId": "a8223c2af342cc0bdc39ed5f82e2f47e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "price",
                                    "category_id",
                                    "instructor_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Laravel cơ bản"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Khóa học Laravel từ A-Z"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 199000
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "instructor_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Khoá học được tạo thành công"
                    }
                }
            }
        },
        "/api/courses/{id}": {
            "get": {
                "tags": [
                    "Course"
                ],
                "summary": "Xem chi tiết khoá học",
                "operationId": "7018863f8e9358f3f1b7e558f544eaae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID khoá học",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin khoá học"
                    }
                }
            },
            "put": {
                "tags": [
                    "Course"
                ],
                "summary": "Cập nhật khoá học",
                "operationId": "f08dedafedb81a971714481fc11aa27e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID khoá học",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Laravel nâng cao"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Chuyên sâu về Laravel"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 299000
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "instructor_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật khoá học thành công"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Course"
                ],
                "summary": "Xóa khoá học",
                "operationId": "8ae94bd55fe6bf987643b506da02e9e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID khoá học",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xoá khoá học thành công"
                    }
                }
            }
        },
        "/api/courses/me/enrolled": {
            "get": {
                "tags": [
                    "Course"
                ],
                "summary": "Lấy danh sách khoá học đã đăng ký của người dùng",
                "operationId": "31eeda8959237dc0d21e132ea7230f3e",
                "responses": {
                    "200": {
                        "description": "Danh sách khoá học đã đăng ký"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/{id}/publish": {
            "post": {
                "tags": [
                    "Course"
                ],
                "summary": "Publish khoá học",
                "operationId": "80a83dc7581de4b09c74f4e47165e29c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID khoá học",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Khoá học đã được publish"
                    }
                }
            }
        },
        "/api/courses/{id}/archive": {
            "post": {
                "tags": [
                    "Course"
                ],
                "summary": "Archive khoá học",
                "operationId": "7e2acf18fa8cee084991911c974b8254",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID khoá học",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Khoá học đã được archive"
                    }
                }
            }
        },
...
    },
...
}